pyproject.toml
[tool.poetry]
name = "discord-bot"
version = "0.1.0"
description = ""
authors = ["0x28F4 <0x28F4@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
py-cord = {extras = ["voice"], version = "^2.6.1"}
google-cloud-speech = "^2.27.0"
retry = "^0.9.2"
google-cloud-texttospeech = "^2.18.0"
openai = "^1.52.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
ruff = "^0.7.1"
pre-commit = "^4.0.1"
mypy = "^1.13.0"
types-pyyaml = "^6.0.12.20240917"
types-retry = "^0.9.9.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
ignore = ["E701", "E501"]
